data-pipelines/test_requirements.txt
langchain-core==0.3.37 # langchain dependencies
langchain-community==0.3.17
langchain-ollama==0.2.3 # for langchain integration with Ollama
langchain-elasticsearch==0.3.2 # for langchain integration with Elastic Search
pypdf==5.3.0
python-dotenv==1.0.1
google-api-python-client==2.157.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pyspark==3.5.4
spark-nlp==5.5.3
ipykernel==6.29.4
sentence-transformers==3.4.1